pyproject.toml
[tool.poetry]
name = "ql-clbf"
version = "0.1.0"
description = ""
authors = ["Daniel Tan <dtch1997@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "ql_clbf"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.10"
pytorch-lightning = {extras = ["extra"], version = "^1.9.4"}
torch = ">=1.3.0"
torchvision = ">=0.6.0"
wandb = "^0.13.11"
setuptools = "65.7.0"
d3rlpy = "^1.1.1"
tensorboard = "^2.12.0"
stable-baselines3 = "1.2.0"
gymnasium = {version = "^0.28.1", extras = ["classic-control", "mujoco"]}
skrl = "^0.10.2"
moviepy = "^1.0.3"
cvxpylayers = "^0.1.5"
freia = "^0.2"
opencv-python = "^4.7.0.72"
mujoco-py = "^2.1.2.14"
d4rl = "^1.1"
mjrl = { git = "https://github.com/aravindr93/mjrl.git"}
minigrid = "<3.0.0"
pandas = "^2.0.1"

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
coverage = "^7.2.1"
pytest = ">=3.0.5"
pytest-cov = "^4.0.0"
pytest-flake8 = "^1.1.1"
flake8 = "^6.0.0"
check-manifest = "^0.49"
twine = "1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
